Make @TestSecurity work correctly with unannotated JAX-RS endpoints security feature
The use of `quarkus.security.jaxrs.deny-unannotated-endpoints=true` essentially
results in the addition of a `DenyAllInterceptor` to the invocation chain
of a JAX-RS endpoint.
Because this interceptor did not take into account the `AuthorizationController`
(like the `RolesAllowedInterceptor` already does), it would result in endpoints
being secured even though security was supposed to be disabled for the specific test.

Fixes: #19896
(cherry picked from commit b9359bf26a866ed186d54ba07c7e29fa2cb38188)