Fix gdb debugging with GraalVM/Mandrel >=21.1
As if
21.1 (https://github.com/oracle/graal/commit/dee66ca6e28fd519cef061130d20e28ec35d1a1f)
dwarf info contain absolute paths to the source-cache. As a result
moving the sources breaks gdb.

The move was performed to make sure the sources are next to the binary,
in an effort to ease debugging on remote hosts by copying the binary and
the sources in the same directory.  As discussed in
https://github.com/oracle/graal/pull/3828 this is not expected to be the
common case, so it's better to favour the local debugging experience by
leaving the sources where they were placed by GraalVM/Mandrel.

(cherry picked from commit db5ad51e9e9178743bf75bafdf2e5a070c31fb4b)