Avoid using Vert.x HttpClient in Dev UI tests
Dev UI tests are dev mode tests which are initiated in the app class
loader, which is the parent of the CL that is supposed to load Vert.x
for the application.
It is not possible to load Vert.x both for the parent class loader and
the children class loaders and the only reason this is working is that
we are leaking class loaders like crazy.

The problem I experienced in my CL cleanup PR was the following:
- Test 1 doesn't load Vert.x at the app level, but Vert.x is loaded in
  the application so with a child QuarkusClassLoader.
- Test 2 is trying to call sendRequest() from the app class loader and
  given we now clean up the child QuarkusClassLoader, it tries to load
  the class from the app CL resulting in a LinkageError.

The JDK HttpClient, while imperfect, allows us to work around this
issue.
Also we were instantiation a new Vert.x instance for each request which
was less than ideal.

Unfortunately, while this problem was responsible of most of the
failures, there are a few others lurking around :/.