Avoid iterating on the resources to get the manifest
We already know from which resource we were able to load the class so
that's the resource in which the package is defined.
In rare occasions, when we have split packages, we might have corner
cases but I don't really expect having a null manifest in one of the
artifact would be very common so that getting to the other might be
beneficial.