Quarkus REST - Be more permissive when adding headers
The JAX-RS API is very loose and allows adding Object and in some cases
we actually add a List to the headers and we need to take this into
account.

While I'm usually not for handling things this way, the API is not
enforcing things properly so I think we really need to handle this
ourself.

Fixes #47507